requirements.txt
Django==4.2.16
python-decouple==3.8
qrcode[pil]==7.4.2
openpyxl==3.1.2
Pillow==10.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==5.0.0
django-redis==5.3.0
whitenoise==6.5.0
django-cors-headers==4.3.1